use a more accurate type in toSize

+++ ThrEd4/wrappers.h
@@ -278,7 +278,7 @@ auto toFloat(inType invar) noexcept(!(std::is_same_v<inType, float> ||
 }
 
 template <class inType>
-auto toSize(inType invar) noexcept(!(std::is_signed_v<inType> || (std::is_same_v<inType, unsigned long long> && std::is_same_v<size_t, uint32_t>)))
+auto toSize(inType invar) noexcept(!(std::is_signed_v<inType> || (std::is_same_v<inType, uint64_t> && std::is_same_v<size_t, uint32_t>)))
     -> size_t {
   if constexpr (std::is_signed_v<inType>) {
 	return gsl::narrow<size_t>(invar);
@@ -288,7 +288,7 @@ auto toSize(inType invar) noexcept(!(std::is_signed_v<inType> || (std::is_same_v
 	  return invar;
 	}
 	else {
-	  if constexpr (std::is_same_v<inType, unsigned long long> && std::is_same_v<size_t, uint32_t>) {
+	  if constexpr (std::is_same_v<inType, uint64_t> && std::is_same_v<size_t, uint32_t>) {
 		return gsl::narrow<size_t>(invar);
 	  }
 	  else {
